README: point to `:nightly` image
The latest one is outdated and the nightly one has been recommended to
me.  Others seemed to be unaware of the nightly one as well, so I
propose to mention it for now.

It can certainly be changed in the future but for now I want to avoid
others falling into the trap of using an old image.

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>