README: add gif to show installation
To help users install the extension, add a gif that shows the workflow.
Also add a small description of the extension to the top.

I want to add more gifs in the future but at the moment there are many
things in motion, so I prefer to wait a couple of days.

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>