update dashboard description
In preparation of the 0.3.0 release, I want to update the dashboard.
At the present, it displays the contents of the README with installation
instructions which are likely confusing users since the extension has
already been installed.

The new text is not perfect and does not fully address #624 but I think
it's a step in the right direction.  The descriptions will very likely
need polishing in the future.  I did not find a way to create links that
allow for opening another menu.  For instance, it would be nice to have
clickable links/buttons that navigate users directly into the specific
menu such as the Recipes Catalog.

Some graphical elements (e.g., screenshots) may be a nice additions as
well.  We can add them later on, if needed, when the UI has settled a
bit more.

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>