chore: remove enhancement template
GitHub defaults to having a feature issue type and no enhancement type, and we're
going to stick with this. This PR removes the enhancement template so that no new
ones will be created. (Updating the feature template was handled separately)

Existing issues with the kind/enhancement label are unaffected.

Part of #3851.

Signed-off-by: Tim deBoer <git@tdeboer.ca>